Выполнил вроде задание, но у меня куда то пропали значения из массива по ключу status, где справа было указано выполнена ли задача Да/Нет
