+++ error.php
@@ -1,0 +1,5 @@
+﻿<?php
+
+echo "Ошибка! Запрашиваемого файла не существует!";
+
+?>
