+++ functions.php
@@ -1,14 +1,23 @@
 <?php 
-
-function get_template($template, $massiv = array()) {
+//функция шаблонизатор
+function get_template($template, $massiv) {
 	if (file_exists($template) == true) {
 		ob_start();
-		require '$template';
-	  extract($massiv = array(), EXTR_PREFIX_ALL);
+		extract ($massiv);
+		require $template;
+	  
 		return ob_get_clean();
 	}
-	else {
-		exit("Ошибка");
+	elseif (file_exists($template) == false) {
+		require_once ("error.php");
 	}
+	
 }
 
+//функция фильтрации данных
+function esc($str) {
+	//$text = htmlspecialchars($str);
+	$text = strip_tags($str);
+
+	return $text;
+}
